Update internal typings and fix auto generated ones
