+++ react/ProductDispatchContext.ts
@@ -1,5 +1,7 @@
 import { createContext, useContext, Dispatch } from 'react'
 
+import { Actions } from './ProductContextProvider'
+
 export const ProductDispatchContext = createContext<Dispatch<Actions> | null>(
   null
 )
